refactor: rename from log.py to logging.py
